update channel type

+++ bot.py
@@ -1424,7 +1424,8 @@ async def bottle(ctx, target_player, bet: str):
                     player_reply = await bot.wait_for(event="message", check=author_check(ctx.author), timeout=30.0)
 
                 player_bet = int(player_reply.content)
-                await player_reply.delete()
+                if not player_reply.channel.type == 'dm':
+                    await player_reply.delete()
                 
                 new_embed = discord.Embed(title='Bottle Game', color=discord.Color.random())
                 new_embed.add_field(name=f"{user.name} Balance",
@@ -1443,7 +1444,8 @@ async def bottle(ctx, target_player, bet: str):
                     challenge_player_reply = await bot.wait_for(event="message", check=author_check(member), timeout=30.0)
                 
                 challenge_player_bet = int(challenge_player_reply.content)
-                await challenge_player_reply.delete()
+                if not challenge_player_reply.channel.type == 'dm':
+                    await challenge_player_reply.delete()
 
                 new_embed = discord.Embed(title='Bottle Game', color=discord.Color.random())
                 new_embed.add_field(name=f"{user.name} Balance",
